joined tree test-files
